+++ Modules/Intent.Modules.Blazor.HttpClients/Templates/TemplateExtensions.cs
@@ -6,6 +6,7 @@ using Intent.Modules.Blazor.HttpClients.Templates.HttpClientConfiguration;
 using Intent.Modules.Blazor.HttpClients.Templates.HttpClientRequestException;
 using Intent.Modules.Blazor.HttpClients.Templates.JsonResponse;
 using Intent.Modules.Blazor.HttpClients.Templates.PagedResult;
+using Intent.Modules.Blazor.HttpClients.Templates.ProblemDetailsWithErrors;
 using Intent.Modules.Blazor.HttpClients.Templates.ServiceContract;
 using Intent.Modules.Common.Templates;
 using Intent.RoslynWeaver.Attributes;
@@ -65,6 +66,11 @@ namespace Intent.Modules.Blazor.HttpClients.Templates
         public static string GetPagedResultName(this IIntentTemplate template)
         {
             return template.GetTypeName(PagedResultTemplate.TemplateId);
+        }
+
+        public static string GetProblemDetailsWithErrorsName(this IIntentTemplate template)
+        {
+            return template.GetTypeName(ProblemDetailsWithErrorsTemplate.TemplateId);
         }
 
         public static string GetServiceContractName<T>(this IIntentTemplate<T> template) where T : Intent.Modelers.Types.ServiceProxies.Api.ServiceProxyModel