+++ Modules/Intent.Modules.Dapr.AspNetCore.ServiceInvocation/release-notes.md
@@ -1,6 +1,7 @@
 ﻿### Version 2.2.2
 
 - Improvement: Added support for Dapr Http Client URL configuration.
+- Improvement: Added support for deserializing ProblemDetails for client HTTP calls.
 
 ### Version 2.2.1